+++ helpers/download.py
@@ -7,18 +7,19 @@
 
 
 # Import libraries
-from urllib.request import urlopen
+from urllib import request
 
 # Import script modules
 from helpers.clean import clean_lines
 
 
-def download_file(url:str) -> dict:
+def download_file(url:str, content_type:str = 'text/html') -> dict:
     '''
     Retrieves a file from a URL and returns the content
 
         Parameters:
             url (str): URL to download the file from
+            content_type (str): optional content type
 
         Returns:
             dict: Provides 'file_type', 'file_extension' and 'content' of the retrieved file
@@ -26,7 +27,9 @@ def download_file(url:str) -> dict:
 
     # Retrieve URL content
     try:
-        response = urlopen(url)
+        request_header = { 'Content-Type': content_type }
+        request_object = request.Request(url, headers = request_header)
+        response = request.urlopen(request_object)
 
         # Check if response is invalid
         if response.status != 200: