+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - Add code of conduct
 - Use speaking command-line arguments
 - Add option to filter resource downloads by string
+- Add optional content negotiation
 - Test everything against the CVMA API